+++ src/test/groovy/fr/marcwrobel/gcli/DatabaseConfigurationTest.groovy
@@ -1,27 +1,29 @@
 package fr.marcwrobel.gcli
 
 import groovy.sql.Sql
-import org.junit.Test
+import org.junit.jupiter.api.Test
 
 import java.sql.SQLException
 
-import static org.junit.Assert.*
+import static org.junit.jupiter.api.Assertions.*
 
 class DatabaseConfigurationTest {
 
   private static final CliTexts TEXTS = new SimpleCliTexts('test')
 
-  @Test(expected = SQLException.class)
+  @Test
   void parsingTest() {
     DatabaseConfiguration configuration = new DatabaseConfiguration(TEXTS)
-    assertTrue(configuration.parse(['-db', 'toto', '-du', 'titi'] as String[]))
 
+    assertTrue(configuration.parse(['-db', 'toto', '-du', 'titi'] as String[]))
     assertEquals('toto', configuration.databaseName)
     assertEquals('titi', configuration.databaseUser)
 
-    configuration.withSql { Sql sql ->
-      println('Hello world !')
-    }
+    assertThrows(SQLException.class, { ->
+      configuration.withSql { Sql sql ->
+        println('Hello world !')
+      }
+    })
   }
 
   @Test