+++ src/test/groovy/fr/marcwrobel/gcli/LogModeTest.groovy
@@ -1,17 +1,18 @@
 package fr.marcwrobel.gcli
 
-import org.junit.Assert
-import org.junit.Test
+import org.junit.jupiter.api.Test
+
+import static org.junit.jupiter.api.Assertions.*
 
 class LogModeTest {
 
   @Test
   void safeValueOfIsCaseInsensitive() {
-    Assert.assertEquals(LogMode.CONSOLE, LogMode.safeValueOf(LogMode.CONSOLE.name().toLowerCase()))
+    assertEquals(LogMode.CONSOLE, LogMode.safeValueOf(LogMode.CONSOLE.name().toLowerCase()))
   }
 
   @Test
   void safeValueOfFallbackToDefaultValue() {
-    Assert.assertEquals(LogMode.DEFAULT, LogMode.safeValueOf(null))
+    assertEquals(LogMode.DEFAULT, LogMode.safeValueOf(null))
   }
 }