+++ src/test/groovy/fr/marcwrobel/gcli/LoggingConfigurationTest.groovy
@@ -1,32 +1,32 @@
 package fr.marcwrobel.gcli
 
 import ch.qos.logback.classic.Level
-import org.junit.Rule
-import org.junit.Test
-import org.junit.rules.TemporaryFolder
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.io.TempDir
 
-import static org.junit.Assert.*
+import static org.junit.jupiter.api.Assertions.*
 
 class LoggingConfigurationTest {
 
   private static final CliTexts TEXTS = new SimpleCliTexts('test')
 
-  @Rule
-  public TemporaryFolder tempFolder = new TemporaryFolder()
+  @TempDir
+  protected File tempFolder;
 
   @Test
   void parsingTest() {
     LoggingConfiguration configuration = new LoggingConfiguration(TEXTS)
-    assertTrue(configuration.parse(['-lm', 'file', '-ll', 'debug', '-ld', tempFolder.root.absolutePath] as String[]))
 
+    assertTrue(configuration.parse(['-lm', 'file', '-ll', 'debug', '-ld', tempFolder.absolutePath] as String[]))
     assertEquals(LogMode.FILE, configuration.logMode)
     assertEquals(Level.DEBUG, configuration.logLevel)
-    assertEquals(tempFolder.root, configuration.logDirectory)
+    assertEquals(tempFolder, configuration.logDirectory)
   }
 
   @Test
   void invalidArgumentTest() {
     LoggingConfiguration configuration = new LoggingConfiguration(TEXTS)
+
     assertFalse(configuration.parse(['-i'] as String[]))
   }